port to linux disabled not ported code
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@24 37977431-3df6-0310-b722-df95706aa16b

+++ rts/AdvTreeGenerator.cpp
@@ -102,7 +102,7 @@ CAdvTreeGenerator::CAdvTreeGenerator()
 	treeNSVP=LoadVertexProgram("treeNS.vp");
 
 	leafDL=glGenLists(8);
-	srand(15);
+	srand((unsigned int)15);
 	for(int a=0;a<8;++a){
 		va=GetVertexArray();
 		va->Initialize();
@@ -119,7 +119,7 @@ CAdvTreeGenerator::CAdvTreeGenerator()
 	}
 
 	pineDL=glGenLists(8);
-	srand(15);
+	srand((unsigned int)15);
 	for(int a=0;a<8;++a){
 		va=GetVertexArray();
 		va->Initialize();
